small change when notifying user of skipped hist.
If the history is skipped in the overview page tt gives the user a hint how to display all history entry (by providing the command to be executed).
This change will take the actual command typed by the user into account to provide this suggestion (use s or show depending of the version the user used).