Support for pause/resume and log
"pause" will pause a currently running tracker
"resume" will resume a tracker for the last_project in the meta section
of the data file. This command does not only work when in pause mode but
will also start a new tracker for last_project if there is no active
tracker.

"log" will display the history of all actions in the past.

Both functions can be disabled in the config section of the data file
but there is no config switch yet to enable or disable either function
from the command line.

The data file format has been extended to deal with the new functions.
There is also a list for log entries that are already written but
functionality to display these logs has not been implemented yet.