Do not sort log before printing
The log is a list of strings in a certain order anyway. It makes no
sense to sort it before printing out to the console. Moreover, the
sorted() function could reorder the action logged at the same time
(e.g. when using "tt switch") and bring them into an incorrect order.