Status will only display total time tracked if necessary
The recently added line showing the total time tracked today in a
particular project (when "status" is called) is only displayed if this
total tracked time differs from the current active tracked time, i.e. if
there is a time slot tracked earlier this day (e.g. if a pause was made)