Updated data file initialization process
Simplyfied the rather error-prone process of initializing a new data file. Now we create a very simple version of the data structure that we initially used when starting this tt project. Thereafter we apply the updates defined in the data file upgrade function.
This removed double code and makes it more maintainable.