changed (default) behaviour when adding a pause
When adding a pause by "pause add …" the time of
the pause will now reduce the recorded time by default.
If you want to keep the recorded hours you must now provide the keyword „append“ to the "pause add …" command.

This has also some effect on the format of the specified time period. You not must use the "T" delimiter between date and time. In former versions a space character (instead of "T") was also recognized. This is not longer the case.